requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==4.2.7
django-ckeditor==6.7.0
django-filter==23.3
django-js-asset==2.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
environs==9.5.0
inflection==0.5.1
marshmallow==3.20.1
packaging==23.2
Pillow==10.1.0
PyJWT==2.8.0
pymemcache==4.0.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
